+++ src/sentimeter/__init__.py
@@ -1,3 +1,3 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-__version__ = "0.0.2"
+__version__ = "0.0.3"
