+++ src/sentimeter/backend.py
@@ -1,0 +1,6 @@
+class Basebackend:
+    def __init__(self, backend_name) -> None:
+        self.backend_name = backend_name
+
+    def process(self, text):
+        return {}
