+++ test_pip_module.py
@@ -1,6 +1,5 @@
 from sentimeter.sentimeter import Sentimeter
 from sentimeter.sources import TextSource
-from sentimeter.backends import AIRemoteBackend
 
 app = Sentimeter()
 source = TextSource(app)
